Formatting fixes and final blog
