+++ package.json
@@ -15,7 +15,7 @@
     "@next/mdx": "^15.3.1",
     "@types/mdx": "^2.0.13",
     "gray-matter": "^4.0.3",
-    "next": "^16.0.3",
+    "next": "^16.1.1",
     "next-mdx-remote-client": "^2.1.2",
     "next-themes": "^0.4.6",
     "react": "^19.2.0",
