+++ projects/doubt-horizon/doubt-horizon.mdx
@@ -10,6 +10,8 @@ addition to learning about technical and game fundamentals, the final project wa
 produced by a team of six. Currently hosted on itch.io,
 [our final game, Doubt's Horizon, is a 2D semi-openworld where you must navigate through the wastes and encounter its inhabitants, some friendly, some fearsome.](https://kaaden.itch.io/doubts-horizon)
 
+<ItchIframe src="https://itch.io/embed/3321970?bg_color=fb922b&amp;fg_color=222222&amp;link_color=630b5c&amp;border_color=333333" width="552" height="167"/>
+
 ![Home screen of Doubt's Horizon](./Screenshot_20250704_183415.avif)
 ![Party on the bridge](./Screenshot_20250704_154957.avif)
