--- projects/game-engine-350/itch-iframe.tsx
@@ -1,29 +1,0 @@
-"use client";
-import { useEffect, useState } from "react";
-
-interface ItchIframeProps {
-  src: string;
-  width: number;
-  height: number;
-}
-
-export function ItchIframe({ src, width, height }: ItchIframeProps) {
-    const [isClient, setIsClient] = useState(false);
-
-  // Temporary hack for client side hydration
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  if (!isClient) {
-    return null;  // Return nothing on the server side
-  }
-
-  return (
-    <div className="flex justify-center items-center mb-6">
-      <iframe src={src} width={width} height={height}>
-        <a href={src}>Link</a>
-      </iframe>
-    </div>
-  );
-}
