+++ src/components/blogpost.tsx
@@ -42,7 +42,7 @@ export default async function BlogPost({ post }: BlogPostProps) {
 
   return (
     <div className="mx-auto my-8 px-6 lg:pl-16 lg:pr-8 lg:flex lg:gap-6">
-      <article className="flex-1">
+      <article className="flex-1 max-w-7xl">
         <h1 className="font-extrabold text-5xl mb-6">{post.frontMatter.title}</h1>
 
         <div className="mx-auto flex flex-row justify-between mb-6 gap-x-2">
@@ -62,7 +62,7 @@ export default async function BlogPost({ post }: BlogPostProps) {
         <MDXRemote source={post.content} components={components} options={options} />
       </article>
 
-      <TableOfContents initialWidth={160} />
+      <TableOfContents initialWidth={200} />
     </div>
   );
 }
